Fixed "skip" on last entity behavior.
Changed the nextEntity method to check if there is a next entity and return BOOL YES if so, No otherwise.  Updated the "Skip Phrase" handling in practice, verify, and entry to use this return value to determine whether to go to the memorize screen for the next entity or the recall screen.

The entry screen also uses this method and return value as part of standard done button handling instead of checking the value explicitly here.

This should also fix an unreported bug of the last entity not having summary rollup information logged.